ejercicio 3 p2

+++ Arboles/include/practicas/p1-3.hpp
@@ -5,12 +5,12 @@ enum operador { NULO, SUMA, RESTA, PROD, DIVI };
 struct elem
 {
     operador o;
-    int n;
+    double n;
     elem(operador o) : o{ o }, n{ 0 } {}
-    elem(int n) : o{ NULO }, n{ n } {}
+    elem(double n) : o{ NULO }, n{ n } {}
 };
 
-int ExpresionInfija(const AbinEnla<elem>& A);
-int ExpresionInfija_REC(AbinEnla<elem>::nodo n, const AbinEnla<elem>& A);
+double ExpresionInfija(const AbinEnla<elem>& A);
+double ExpresionInfija_REC(AbinEnla<elem>::nodo n, const AbinEnla<elem>& A);
 
 #endif
